Created a test lib for api tests.
