+++ api/handler/spaces_resource_blackbox_test.go
@@ -8,6 +8,7 @@ import (
 	. "github.com/onsi/ginkgo"
 
 	"github.com/fabric8-services/fabric8-wit/api/model"
+	. "github.com/fabric8-services/fabric8-wit/api/test"
 	"github.com/fabric8-services/fabric8-wit/gormsupport/cleaner"
 	"github.com/fabric8-services/fabric8-wit/gormtestsupport"
 	"github.com/google/jsonapi"
